+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 * Define `UT` joint scheme
 * Define joint scheme for `Maybe` datatype
 * Define joint scheme for `Either` datatype
+* Define joint scheme for `Reader` datatype